+++ tests/test_solve_ik.py
@@ -19,30 +19,81 @@
 Test inverse kinematics.
 """
 
+import os
 import unittest
 
 import numpy as np
 
+from pink import apply_configuration
 from pink import solve_ik
-from pink.models import build_jvrc_model
+from pink.models import build_from_urdf
+from pink.tasks import BodyTask
 
 
 class TestSolveIK(unittest.TestCase):
+
+    """
+    Test fixture for the solve_ik function.
+    """
+
     def setUp(self):
-        """
-        Prepare test fixture.
-        """
-        self.robot = build_jvrc_model()
-        self.assertIsNotNone(self.robot)
+        models_dir = os.path.join(os.path.dirname(__file__), "models")
+        print(models_dir, "===")
+        jvrc_description = os.path.join(models_dir, "jvrc_description")
+        upkie_description = os.path.join(models_dir, "upkie_description")
+        self.jvrc_description = jvrc_description
+        self.upkie_description = upkie_description
 
     def test_no_task(self):
         """
         Raise an error when the robot body is not found.
         """
+        robot = build_from_urdf(self.jvrc_description)
         tasks = []
-        dt = 1e-3  # [s]
-        v = solve_ik(self.robot, tasks, dt)
-        self.assertTrue(np.allclose(v, np.zeros(self.robot.nv)))
+        v = solve_ik(robot, tasks, dt=1e-3)
+        self.assertTrue(np.allclose(v, np.zeros(robot.nv)))
+
+    def test_single_task_fulfilled(self):
+        robot = build_from_urdf(self.upkie_description)
+        configuration = apply_configuration(robot, robot.q0)
+        left_contact_task = BodyTask(
+            "left_contact", position_cost=1.0, orientation_cost=1.0
+        )
+        left_contact_task.set_target(
+            configuration.get_transform_body_to_world("left_contact")
+        )
+        velocity = solve_ik(configuration, [left_contact_task], dt=5e-3)
+        self.assertTrue(np.allclose(velocity, 0.0))
+
+    def test_three_tasks_fulfilled(self):
+        """
+        No motion when all targets are reached.
+        """
+        robot = build_from_urdf(self.jvrc_description)
+        configuration = apply_configuration(robot, robot.q0)
+        left_ankle_task = BodyTask(
+            "l_ankle", position_cost=1.0, orientation_cost=3.0
+        )
+        right_ankle_task = BodyTask(
+            "r_ankle", position_cost=1.0, orientation_cost=3.0
+        )
+        pelvis_task = BodyTask(
+            "PELVIS_S", position_cost=1.0, orientation_cost=3.0
+        )
+
+        left_ankle_task.set_target(
+            configuration.get_transform_body_to_world("l_ankle")
+        )
+        right_ankle_task.set_target(
+            configuration.get_transform_body_to_world("r_ankle")
+        )
+        pelvis_task.set_target(
+            configuration.get_transform_body_to_world("PELVIS_S")
+        )
+
+        tasks = [pelvis_task, left_ankle_task, right_ankle_task]
+        velocity = solve_ik(configuration, tasks, dt=5e-3)
+        self.assertTrue(np.allclose(velocity, 0.0))
 
 
 if __name__ == "__main__":